% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_repos.R
\name{browse_repos}
\alias{browse_repos}
\title{Browse repositories contained in a list}
\usage{
browse_repos(db, idx)
}
\arguments{
\item{db}{A list with the same structure as \link{.gisrepos}.}

\item{idx}{An index (optional) passed to \code{\link[=subset_db]{subset_db()}}.}
}
\description{
All links contained in a list with the structure of \link{.gisrepos} will be
opened in the default browser.
}
\examples{
browse_repos(.gisrepos, 1:2)

}
