% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to PostgreSQL database}
\usage{
connect_db(
  dbname = "",
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  package = c("RPostgres", "RPostgreSQL"),
  ...
)
}
\arguments{
\item{dbname, host, port, user, password}{Each a character value passed to
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}. They can also be modified in the prompt.}

\item{package}{A character value indicating the name of the package to be
used for connecting the database (either "RPostgres" or "RPostgreSQL").}

\item{...}{Further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
This function opens a prompt to insert connection details to a PostgreSQL
database. Alternatively the package \code{RPostgres} or \code{RPostgreSQL} can be used
(see argument \code{package}).
}
\references{
\preformatted{http://r.789695.n4.nabble.com/tkentry-that-exits-after-RETURN-tt854721.html#none
https://gist.github.com/mages/2aed2a053e355e3bfe7c#file-getlogindetails-r
Dalgaard (2001)
}
}
\author{
Miguel Alvarez
}
