% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2df.R
\name{data2df}
\alias{data2df}
\title{Convert data set list into data.frame}
\usage{
data2df(x, sep = ";\\n")
}
\arguments{
\item{x}{A list with the structure of object \code{.gisrepos}.}
}
\description{
This function creates a data.frame with the content of a list. Relations
associated to the main table will be pasted and separated by semicolons (e.g.
links or keywords).

There is also a counter-function \code{df2data()}.
}
\keyword{internal}
