% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2text.R
\name{data2text}
\alias{data2text}
\title{Merge descriptive columns (text) in a single string}
\usage{
data2text(x)
}
\description{
This function makes descriptive columns suitable for search content by using
function \code{\link[=grepl]{grepl()}}.
}
\keyword{internal}
