% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisrepos-data.R
\docType{data}
\name{.gisrepos}
\alias{.gisrepos}
\title{List of GIS repositories}
\format{
A list.
}
\description{
The basic data set is contained in a list including related data frames.
In this list the main information is contained in \code{data_sets}.
}
\examples{
str(.gisrepos)

}
