% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr2list.R
\name{gr2list}
\alias{gr2list}
\alias{gr2list,gisrepos-method}
\title{Coerce gisrepos objects into list}
\usage{
\S4method{gr2list}{gisrepos}(gr, ...)
}
\arguments{
\item{gr}{An object of class \linkS4class{gisrepos}.}

\item{...}{Further arguments passed among methods.}
}
\description{
For convenience repositories will be organized in a list format, where each
element in the list corresponds to a repository entry.
}
