% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template.R
\name{read_template}
\alias{read_template}
\title{Read the updated template}
\usage{
read_template(file, main_table = "main_table", ...)
}
\arguments{
\item{file}{A character vector indicating the path and the name of the
template Excel book.}

\item{main_table}{A character value indicating the name of the sheet
containing the main table in the Excel book.}

\item{...}{Further arguments passed to \code{\link[=read_ods]{read_ods()}}.}
}
\description{
This function imports the updated entries that are stored in a standard
Excel sheet into a data frame in the running R session.
}
\keyword{internal}
