% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html.R
\name{render_html}
\alias{render_html}
\title{Render data set in an html display}
\usage{
render_html(
  filename = ".temp",
  db = .gisrepos,
  title = "Spatial Repositories",
  author = "Alvarez et al.",
  date_format = "\%d-\%m-\%Y",
  keep_rmd = FALSE
)
}
\arguments{
\item{filename}{Character value indicating the name of the output html file
(without file extension).}

\item{db}{A list with the structure of object '.gisrepos'.}

\item{title}{Character value with the title for the document.}

\item{author}{Character value with the name of the document's author.}

\item{date_format}{Format of date information (passed to \code{\link[=format]{format()}}).}

\item{keep_rmd}{Logical value indicating whether the produced Rmd file
should be kept in the working directory or deleted.}
}
\description{
Function required to produce an overview in the browser
}
\keyword{internal}
