% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_now.R
\name{require_now}
\alias{require_now}
\title{Install packages only if not yet installed}
\usage{
require_now(pkgs, update = FALSE, ...)
}
\arguments{
\item{pkgs}{Character vector, the names of the packages to be installed.}

\item{update}{Logical value indicating whether installed packages should be
updated or not (done by \code{\link[=update.packages]{update.packages()}}).}

\item{...}{Further arguments passed to \code{\link[=install.packages]{install.packages()}}.}
}
\description{
A list of packages will be compared with the installed ones and if missing,
installed. These packages will be also loaded in the current session.
This function is a wrapper of \code{\link[=update.packages]{update.packages()}} and \code{\link[=install.packages]{install.packages()}}.
}
\author{
Markus Alfter and Miguel Alvarez.
}
