% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_db.R
\name{subset_db}
\alias{subset_db}
\title{Produce a subset of repository list}
\usage{
subset_db(db, idx)
}
\arguments{
\item{db}{A list structured in the same way as \link{.gisrepos}.}

\item{idx}{A vector indicating the rows to be selected in the data frame
\code{data_sets} within 'db'. It may be either an integer, a logical or a
character vector. In the later case, rows in \code{data_sets} have to be
named.}
}
\value{
A list including the subset of 'db'.
}
\description{
Produces a subset considering all relationships between data frames included
in a list structured as in \link{.gisrepos}.
}
\examples{
first_repos <- subset_db(.gisrepos, 1:5)

}
