% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary,gisrepos-method}
\alias{print,gisrepos-method}
\title{Summary of gisrepos object}
\usage{
\S4method{summary}{gisrepos}(object, ...)

\S4method{print}{gisrepos}(x, ...)
}
\arguments{
\item{object, x}{A \linkS4class{gisrepos} object.}

\item{...}{Further arguments passed among methods.}
}
\description{
A brief overview of the content of the object. This overview will be also
printed in the console when calling it as a symbol.
}
