% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.R
\name{view_html}
\alias{view_html}
\title{Display rendered html file}
\usage{
view_html(filename)
}
\arguments{
\item{filename}{Character vector showing the name of the file to be
displayed.}
}
\description{
This function displays files rendered by \code{\link[=render]{render()}} or \code{\link[=render_html]{render_html()}} in the
default browser.
}
\keyword{internal}
