% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_repo.R
\name{search_repo}
\alias{search_repo}
\title{Search GIS repositories}
\usage{
search_repo(topic, db = .gisrepos, filename = ".temp", ...)
}
\arguments{
\item{topic}{A character value to be matched in the content of the database.}

\item{db}{Database used for search (by default pre-installed '.gisrepos').}

\item{filename}{A character value with the name of the ouput html file
(without file extension).}

\item{...}{Further arguments.}
}
\description{
This function retrieves a display of data sets matching a 'topic' by using
\code{\link[=grepl]{grepl()}} with option \code{ignore.case=FALSE}.
}
\examples{
search_repo("temp")

}
